- Déplace les fichiers de l'interface utilisateur depuis le dossier 'ml_logic' vers le dossier 'interface' - Améliore la lisibilité du formulaire - Ajoute une image de fond à l'interface
