+++ impact_carbone/interface/ui_files/.streamlit/config.toml
@@ -1,0 +1,12 @@
+[theme]
+backgroundColor = "white" # ==> PROBLEME DE COULEUR DANS LES COMBOBOX
+secondaryBackgroundColor = "white" # ==> IMPOSSIBLE A MODIFI
+
+[browser]
+gatherUsageStats = false
+
+[client]
+toolbarMode = "minimal"
+
+[ui]
+hideTopBar = true
