+++ impact_carbone/ml_logic/main.py
@@ -24,7 +24,7 @@ with open(model_path, 'rb') as model_file:
 x_new = 0
 
 # Faire la prédiction avec le modèle chargé
-y_pred_new = predict_x(x_new, cf, best_gbr)
+#y_pred_new = predict_x(x_new, cf, best_gbr)
 
 print(f"Votre score de pollution est de {y_pred_new}")
 
@@ -51,15 +51,15 @@ else:
 
 
 # Graphique à retourner sur le Streamlite pour montrer où se trouve l'individu face à son pays
-fig = graphique(y_pred_new,data_country_co2)
-fig.show()
+#fig = graphique(y_pred_new,data_country_co2)
+#fig.show()
 
 # M'occuper du Sharp avec Hana à remplir en output
 # Graphique en camember à retourner pour montrer l'importance des groupes pour la pred
-fig_pie = visualize_shap_pie_by_group(best_gbr,x_new, sample_ind=0)
-fig_pie.show()
+#fig_pie = visualize_shap_pie_by_group(best_gbr,x_new, sample_ind=0)
+#fig_pie.show()
 
 # Recommandations spécialisées à l'utilisateur à afficher sur le streamlite à la fin
-recommendations = generate_recommendations(best_gbr, x_new, sample_ind=0, top_n=3)
-for rec in recommendations:
-    print(rec)
+#recommendations = generate_recommendations(best_gbr, x_new, sample_ind=0, top_n=3)
+#for rec in recommendations:
+#    print(rec)
