--- impact_carbone/ml_logic/ui_files/default_values.py
@@ -1,24 +1,0 @@
-from ui_files.constants import DICT_DISPLAY_TO_MDL
-
-body_type = DICT_DISPLAY_TO_MDL['body_type'][0][0]
-gender = DICT_DISPLAY_TO_MDL['gender'][0][0]
-diet = DICT_DISPLAY_TO_MDL['diet'][0][1]
-
-waste_bag_weekly_count = 2
-waste_bag_type = DICT_DISPLAY_TO_MDL['waste_bag_type'][0][1]
-recycling = DICT_DISPLAY_TO_MDL['recycling'][0]
-personal_hygiene = DICT_DISPLAY_TO_MDL['personal_hygiene'][0][1]
-heating = DICT_DISPLAY_TO_MDL['heating'][0][3]
-cooking = [DICT_DISPLAY_TO_MDL['cooking'][0][1]]
-efficiency = DICT_DISPLAY_TO_MDL['efficiency'][0][2]
-
-transportation_mode = DICT_DISPLAY_TO_MDL['transportation_mode'][0][0]
-energy_source = DICT_DISPLAY_TO_MDL['energy_source'][0][0]
-monthly_distance = 900
-air_travel = DICT_DISPLAY_TO_MDL['air_travel'][0][0]
-
-social_activity = DICT_DISPLAY_TO_MDL['social_activity'][0][2]
-monthly_grocery = 250
-screen_time = 3
-internet_time = 3
-clothes = 2
